+++ src/main/java/com/cotato/kampus/domain/board/application/BoardService.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Service;
 import com.cotato.kampus.domain.board.domain.Board;
 import com.cotato.kampus.domain.board.domain.BoardWithFavoriteStatus;
 import com.cotato.kampus.domain.board.domain.HomeBoardAndPostPreview;
-import com.cotato.kampus.domain.board.implement.BoardDtoEnhancer;
-import com.cotato.kampus.domain.board.implement.BoardFavoriteReader;
-import com.cotato.kampus.domain.board.implement.BoardFinder;
+import com.cotato.kampus.domain.board.implement.board.BoardDtoEnhancer;
+import com.cotato.kampus.domain.board.implement.boardFavorite.BoardFavoriteReader;
+import com.cotato.kampus.domain.board.implement.board.BoardFinder;
 import com.cotato.kampus.domain.common.application.ApiUserResolver;
 import com.cotato.kampus.domain.post.application.PostDtoMapper;
 import com.cotato.kampus.domain.post.application.PostFinder;
