+++ src/main/java/com/cotato/kampus/domain/board/implement/boardFavorite/BoardFavoriteDeleter.java
@@ -1,4 +1,4 @@
-package com.cotato.kampus.domain.board.implement;
+package com.cotato.kampus.domain.board.implement.boardFavorite;
 
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
